(#41): 숏폼에서 선택한 관광지가 가운데 유지되도록 하는 로직 추가
