+++ src/pages/map/aroundSearch/components/GeoAroundTouristMap.tsx
@@ -2,11 +2,16 @@ import { Map } from 'react-kakao-maps-sdk';
 import NearbyTouristAttractionPinPoint from './NearbyTouristAttractionPinpoint';
 import CurrentDeviceLocation from '../../components/CurrentDeviceLocation';
 import { TouristContentsTypeFilter } from '@/components';
-import { useMemo, useRef, useState } from 'react';
-import type { AroundContentTypeId, GeoTripLocation } from '@/pages/types';
+import { useEffect, useMemo, useRef, useState } from 'react';
+import type {
+  AroundContentTypeId,
+  GeoTripLocation,
+  TourItem,
+} from '@/pages/types';
 import useAroundTouristQuery from '../service/getAroundTouristMapData';
 import { withAroundMapParams } from '../../components';
 import { ResizingMap } from '../../destination/components';
+import MiddleContent from './MiddleContent';
 
 interface GeoAroundTouristMapProps {
   location: GeoTripLocation;
@@ -16,7 +21,6 @@ interface GeoAroundTouristMapProps {
 
 function GeoAroundTouristMap({
   location,
-  contentId,
   tourContentTypeId,
 }: GeoAroundTouristMapProps) {
   const [selectedContentTypeId, setSelectedContentTypeId] =
@@ -25,8 +29,15 @@ function GeoAroundTouristMap({
     location,
     selectedContentTypeId
   );
+  const middleTouristRef = useRef<TourItem | null>(null);
 
-  const middleTouristObject = useRef(aroundTouristObjects[0]);
+  const middleTouristObject = useMemo(() => {
+    if (aroundTouristObjects.length > 0 && middleTouristRef.current === null) {
+      middleTouristRef.current = aroundTouristObjects[0];
+      return aroundTouristObjects[0];
+    }
+    return middleTouristRef.current;
+  }, [aroundTouristObjects]);
 
   const allLocation = useMemo(() => {
     return aroundTouristObjects.map(tourist => ({
@@ -53,11 +64,8 @@ function GeoAroundTouristMap({
           />
         );
       })}
-      {aroundTouristObjects.length > 0 && (
-        <NearbyTouristAttractionPinPoint
-          tourist={middleTouristObject.current}
-          location={location}
-        />
+      {aroundTouristObjects.length > 0 && middleTouristObject && (
+        <MiddleContent location={location} tourist={middleTouristObject} />
       )}
     </Map>
   );
