+++ src/pages/map/aroundSearch/components/index.ts
@@ -1,2 +1,3 @@
 export { default as GeoAroundTouristMap } from './GeoAroundTouristMap';
 export { default as NearbyTouristAttractionPinPoint } from './NearbyTouristAttractionPinpoint';
+export { default as AroundContent } from './AroundContent';
